pyproject.toml
[project]
name = "tecd-retail-recsys"
version = "0.1.0"
description = "Neural network-based recommendation system for retail basket predictions using T-ECD dataset"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Danila Lyapin" }]
dependencies = [
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "mlflow>=2.9.0",
    "dvc>=3.30.0",
    "dvc-s3>=3.0.0",
    "fire>=0.5.0",
    "tqdm>=4.66.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
    "pytest>=7.4.0",
]

[project.scripts]
tecd-recsys = "tecd_retail_recsys.commands:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["tecd_retail_recsys"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
